refundqueen/requirements.txt
flask
opencv-python-headless
pillow
pytesseract
requests
beautifulsoup4
gunicorn